ASoC: msm: qdsp6v2: send right stream ID to DSP for FLAC
For flac format, when starting the next track, right stream
ID need to be send to DSP as part of ASM command to update
media format. Unless the new stream ID is sent, DSP will
try to apply the configs to the old stream ID and will
eventually timeout.

CRs-Fixed: 767487
Change-Id: Ie8e3b0d1ee9c819f24b01e6a5eb1f70ea5f7f828
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>