ASoC: msm: qdsp6v2: add support for any number of ASM buffers
Current logic in ASM driver supports CPU and DSP buffers only
when the number of buffers is power of 2. At times, the number
of buffers might need to be different than power of 2. Modify
the logic to find next available CPU or DSP buffer, even when
the number of buffers is anything other than power of 2.

CRs-Fixed: 771446
Change-Id: Ibda6e0843387f6286a75c1bf4dc2068b4c4e7e9b
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>

Signed-off-by: SoUnd001 <sound0
020@gmail.com>